# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controller.CreateCommentRequest:
    properties:
      content:
        example: 写得很好！
        maxLength: 1000
        minLength: 1
        type: string
      parent_id:
        type: integer
    required:
    - content
    type: object
...
  controller.SwaggerCommentListData:
    properties:
      list:
        items:
          $ref: '#/definitions/controller.SwaggerCommentResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  controller.SwaggerCommentResponse:
    properties:
      content:
        example: 写得很好！
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      like_count:
        example: 0
        type: integer
      parent_id:
        description: 移除 example 标签
        type: integer
      post_id:
        example: 1
        type: integer
      replies:
        items:
          $ref: '#/definitions/controller.SwaggerCommentResponse'
        type: array
      status:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      username:
        example: testuser
        type: string
    type: object
...
  utils.Response:
    properties:
      code:
        description: 状态码，0表示成功
        example: 0
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
  description: 个人博客系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
...
  /admin/comments/{cid}/approve:
    post:
      consumes:
      - application/json
      description: 审核通过待审核的评论（需要管理员权限）
      parameters:
      - description: 评论ID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 审核通过
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 审核通过评论
      tags:
      - 评论管理
  /admin/comments/{cid}/reject:
    post:
      consumes:
      - application/json
      description: 拒绝待审核的评论（需要管理员权限）
      parameters:
      - description: 评论ID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 已拒绝
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 拒绝评论
      tags:
      - 评论管理
  /admin/comments/pending:
    get:
      consumes:
      - application/json
      description: 获取所有待审核的评论（需要管理员权限）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCommentListData'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取待审核评论
      tags:
      - 评论管理
...
  /comments/{cid}:
    delete:
      consumes:
      - application/json
      description: 删除评论（需要认证，仅作者或管理员）
      parameters:
      - description: 评论ID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论管理
  /comments/{cid}/like:
    delete:
      consumes:
      - application/json
      description: 取消评论点赞（需要认证）
      parameters:
      - description: 评论ID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 取消点赞成功
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 取消点赞
      tags:
      - 评论管理
    post:
      consumes:
      - application/json
      description: 给评论点赞（需要认证）
      parameters:
      - description: 评论ID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 点赞成功
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 点赞评论
      tags:
      - 评论管理
...
  /posts/{id}/comments:
    get:
      consumes:
      - application/json
      description: 获取指定文章的所有评论（树形结构）
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCommentListData'
              type: object
      summary: 获取文章评论列表
      tags:
      - 评论管理
    post:
      consumes:
      - application/json
      description: 对文章发表评论或回复（需要认证）
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 评论成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCommentResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 发表评论
      tags:
      - 评论管理
...
  /profile/comments:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的所有评论（需要认证）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerCommentListData'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取用户评论列表
      tags:
      - 评论管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"